Add files via upload
What is in here?

Loyn, R.H. (1987) Effects of patch area and habitat on bird abundances, species numbers and tree health in fragmented Victorian forests. In: Nature Conservation: the Role of Remnants of Native Vegetation (Saunders, D.A., Arnold, G.W., Burbidge, A.A.&Hopkins, A.J.M. eds.), pp. 65–77. Surrey Beatty & Sons, Chipping Norton, NSW.

The focus of Loyn (1987) was establishing what characteristics of habitat were related to the abundance and diversity of forest birds. For this exercise, we will focus on 56 selected forest patches in southeastern Victoria (Australia). Loyn (1987) recorded in these patches the abundance of forest birds (`ABUND`) in each patch as a response variable. The predictor variables recorded for each patch included area (ha; `AREA`), the number of years since the patch was isolated by clearing (years; `YR.ISOL`), the distance to the nearest patch (km; `DIST`), the distance to the nearest larger patch (km; `LDIST`), an index of stock grazing history (`GRAZE`) ranging from 1 (light) to 5 (heavy), and mean altitude (m; `ALT`).

The goal of this exercise is to examine this dataset, aiming to understand which aspects of habitat and human activity affect the biodiversity and abundance of organisms within remnant patches of forest.

File contents:
	* ABUND - abundance of forest birds in patch
	* AREA - patch area (ha)
	* YR.ISOL - number of years since patch was isolated
	* DIST - distance to nearest patch (km)
	* LDIST - distance to nearest larger patch (km)
	* GRAZE - grazing intesity in patch (1 to 5 indicating light to heavy)
	* ALT - altitude (m above sealevel) of patch
	* L10DIST - log10 transformation of DIST
	* L10LDIST - log10 transformation of LDIST
	* L10AREA - log10 transformation of LAREA

	* CYR.ISOL - centered YR.ISOL
	* CL10AREA - centered L10AREA
	* CGRAZE - centered GRAZE
	* RESID1 - residuals from linear regression of ABUND against L10AREA + YR.ISOL + GRAZE + L10AREA*YR.ISOL + YR.ISOL*GRAZE + L10AREA*GRAZE + L10AREA*YR.ISOL*GRAZE
	* PREDICT1 - predicted ABUND from linear regression of ABUND against L10AREA + YR.ISOL + GRAZE + L10AREA*YR.ISOL + YR.ISOL*GRAZE + L10AREA*GRAZE + L10AREA*YR.ISOL*GRAZE
	* AREARESY - residuals from OLS regression of ABUND against L10DIST + L10LDIST + YR.ISOL + GRAZE + ALT
	* AREARESX - residuals from OLS regression of L10AREA against L10DIST + L10LDIST + YR.ISOL + GRAZE + ALT
	* GRAZRESY - residuals from OLS regression of ABUND against L10DIST + L10LDIST + YR.ISOL + L10AREA + ALT
	* GRAZRESX - residuals from OLS regression of GRAZE against L10DIST + L10LDIST + YR.ISOL + L10AREA + ALT 
	* YRRESY - residuals from OLS regression of ABUND against L10DIST + L10LDIST + L10AREA + GRAZE + ALT
	* YRRESX - residuals from OLS regression of YR.ISOL against L10DIST + L10LDIST + L10AREA + GRAZE + ALT